refactor: refactored reddit fetching code
created a utils file for all reddit specific functionalities in one file
