+++ new.md
@@ -1,0 +1,1 @@
+I’ve noticed that many foreigners are often confused when they come to Thailand and eat at Thai restaurants. Instead of using a knife and fork like they’re used to, here in Thailand we use a spoon and fork to eat our meals. Has anyone else experienced this kind of confusion or been unsure of what to do? Let me explain why we do it this way and how it works. In Thailand, we usually use a spoon and fork because most of our meals are served with rice, which is our staple food. The spoon helps to scoop the rice and other dishes more easily, while the fork is used to push food onto the spoon. Unlike in some countries where the fork is used for stabbing food, here it's more of a tool to assist the spoon. However, if it’s a steak, we do use a knife and fork. And for noodle dishes, we use chopsticks with a spoon. We choose our utensils depending on the type of food, so it varies with the meal! When we use the spoon and fork, we hold the spoon in our dominant hand, and the fork in the other hand. The fork is used to push food onto the spoon, and then we use the spoon to eat. If you're eating something with sauce or curry, you can directly scoop it up with the spoon. In some countries, people use knives and forks, or even their hands, to eat. But in Thailand, the spoon and fork are the most common utensils, especially for meals that include rice. It’s a convenient and practical way to enjoy our food. ![img](3rjswazwpvrd1)
